+++ packages/fe-fpm-writer/src/controller-extension/index.ts
@@ -1,7 +1,7 @@
 import { create as createStorage } from 'mem-fs';
 import { create } from 'mem-fs-editor';
 import type { Editor } from 'mem-fs-editor';
-import { join } from 'path';
+import { join } from 'node:path';
 import { render } from 'ejs';
 import type {
     ControllerExtension,