+++ packages/fe-fpm-writer/test/unit/app.test.ts
@@ -2,7 +2,7 @@ import type { ManifestNamespace } from '@sap-ux/project-access';
 import { create as createStorage } from 'mem-fs';
 import type { Editor } from 'mem-fs-editor';
 import { create } from 'mem-fs-editor';
-import { join } from 'path';
+import { join } from 'node:path';
 import { enableFPM, MIN_VERSION } from '../../src/app';
 import type { Manifest } from '../../src/common/types';