+++ packages/fiori-annotation-api/src/cds/service.ts
@@ -1,4 +1,4 @@
-import { pathToFileURL } from 'url';
+import { pathToFileURL } from 'node:url';
 
 import { getCdsFiles } from '@sap/ux-cds-compiler-facade';