+++ packages/preview-middleware/src/base/flex.ts
@@ -1,8 +1,8 @@
 import type { Logger } from '@sap-ux/logger';
 import type { ReaderCollection } from '@ui5/fs';
 import type { Editor } from 'mem-fs-editor';
-import { existsSync, readdirSync, statSync, unlinkSync } from 'fs';
-import { join, parse, sep } from 'path';
+import { existsSync, readdirSync, statSync, unlinkSync } from 'node:fs';
+import { join, parse, sep } from 'node:path';
 import type { CommonChangeProperties } from '@sap-ux/adp-tooling';
 
 /**