+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.1.48
+
+### Patch Changes
+
+-   Updated dependencies [6debb89]
+    -   @sap-ux/fe-fpm-writer@0.39.10
+
+## 0.1.47
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.39.9
+
+## 0.1.46
+
+### Patch Changes
+
+-   Updated dependencies [43dc48c]
+    -   @sap-ux/fe-fpm-writer@0.39.8
+
+## 0.1.45
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.39.7
+
+## 0.1.44
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.39.6
+
+## 0.1.43
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.39.5
+
+## 0.1.42
+
+### Patch Changes
+
+-   43a2446: chore: fix Sonar issues
+-   Updated dependencies [43a2446]
+    -   @sap-ux/fe-fpm-writer@0.39.4
+
 ## 0.1.41
 
 ### Patch Changes