+++ examples/ui-prompting-examples/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @sap-ux-private/ui-prompting-examples
 
+## 0.4.5
+
+### Patch Changes
+
+-   Updated dependencies [6debb89]
+-   Updated dependencies [6debb89]
+    -   @sap-ux/ui-components@1.26.13
+    -   @sap-ux/ui-prompting@0.5.4
+
+## 0.4.4
+
+### Patch Changes
+
+-   Updated dependencies [4c448a7]
+    -   @sap-ux/ui-components@1.26.12
+    -   @sap-ux/ui-prompting@0.5.3
+
+## 0.4.3
+
+### Patch Changes
+
+-   43a2446: chore: fix Sonar issues
+-   Updated dependencies [43a2446]
+    -   @sap-ux/ui-prompting@0.5.2
+
 ## 0.4.2
 
 ### Patch Changes