+++ packages/fiori-docs-embeddings/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/fiori-docs-embeddings
 
+## 0.3.0
+
+### Minor Changes
+
+-   e75b594: Add Fiori Development portal documentation to embeddings
+
+## 0.2.0
+
+### Minor Changes
+
+-   b5b6c9c: Add FPM Development portal documentation to embeddings
+
+## 0.1.1
+
+### Patch Changes
+
+-   43a2446: chore: fix Sonar issues
+
 ## 0.1.0
 
 ### Minor Changes